requirements.txt
Django==1.11.2
django-material==1.2.2
djangorestframework==3.7.7
pytz==2018.3
